Windows build tweakage

+++ html/muscle-by-example/examples/networkutilityfunctions/example_1_tcp_client.cpp
@@ -23,9 +23,9 @@ int main(int argc, char ** argv)
    {
       // Send HTTP request out to the Google
       const char httpRequestStr[] = "GET /index.html HTTP/1.0\r\n\r\n";
-      const int reqLen = strlen(httpRequestStr);
-      const int numBytesSent = WriteData(tcpSock, httpRequestStr, reqLen, true);
-      LogTime(MUSCLE_LOG_INFO, "%i/%i bytes of HTTP request data sent to the server.\n", numBytesSent, reqLen);
+      const uint32 reqLen = (uint32) strlen(httpRequestStr);
+      const int32 numBytesSent = WriteData(tcpSock, httpRequestStr, reqLen, true);
+      LogTime(MUSCLE_LOG_INFO, "%i/%u bytes of HTTP request data sent to the server.\n", (int) numBytesSent, (unsigned int) reqLen);
 
       // Read back the server's response data and print it to stdout
       char buf[2048];
